# swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  description: Complete API documentation for E-commerce platform with product categories, brands, models, layout management, role-based access, reCAPTCHA protection, and auto-generated IDs
  version: 4.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.ecommerce.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignupRequest:
      allOf:
        - $ref: '#/components/schemas/RecaptchaToken'
        - type: object
          required:
            - name
            - surname
            - email
            - password
            - phone
          properties:
            name:
              type: string
              example: "John"
              description: "User's first name"
            surname:
              type: string
              example: "Doe"
              description: "User's last name"
            email:
              type: string
              format: email
              example: "john.doe@example.com"
              description: "User's email address (must be unique)"
            password:
              type: string
              format: password
              example: "Password123"
              description: "Password (minimum 6 characters)"
            phone:
              type: string
              example: "9876543210"
              description: "User's phone number"
            addresses:
              type: array
              description: "Optional array of addresses"
              items:
                $ref: '#/components/schemas/Address'
    LoginRequest:
      allOf:
        - $ref: '#/components/schemas/RecaptchaToken'
        - type: object
          required:
            - email
            - password
          properties:
            email:
              type: string
              format: email
              example: "john.doe@example.com"
              description: "User's email address"
            password:
              type: string
              format: password
              example: "Password123"
              description: "User's password"

    # ============= Category Schemas (UPDATED with Brands) =============
...
    CategoryTree:
      type: object
      properties:
        id:
          type: string
          example: "CAT123456781234"
        name:
          type: string
          example: "Electronics"
        slug:
          type: string
          example: "electronics"
        order:
          type: integer
          example: 1
        image:
          type: string
        brands:
          type: array
          items:
            type: string
          example: ["Samsung", "Sony", "LG"]
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTree'
...
    CategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Laptops"
          description: "Category name (will auto-generate slug)"
        description:
          type: string
          example: "Portable computers"
        brands:
          oneOf:
            - type: array
              items:
                type: string
              example: ["Dell", "HP", "Lenovo", "Apple"]
            - type: string
              example: "Dell,HP,Lenovo,Apple"
          description: "Array of brands or comma-separated string"
        parent:
          type: string
          example: "CAT123456781234"
          description: "Parent category ID (can be ObjectId or categoryId string, use '0' for root)"
        image:
          type: string
          example: "https://example.com/laptops.jpg"
        order:
          type: integer
          example: 1
          description: "Display order (default: 0)"
...
    Product:
      type: object
      properties:
        _id:
          type: string
        productId:
          type: string
          example: "PRD123456781234"
          description: "Auto-generated product ID (format: PRD + timestamp + random)"
        name:
          type: string
          example: "Intel Core i9-13900K"
        heading:
          type: string
          example: "24-Core Desktop Processor"
        description:
          type: string
        category:
          $ref: '#/components/schemas/CategoryInfo'
        brand:
          type: string
          example: "Intel"
          description: "Product brand (required)"
        model:
          type: string
          example: "Core i9-13900K"
          description: "Product model (optional)"
        price:
          type: number
          example: 58999
        cost:
          type: number
          example: 52000
        warranty:
          type: string
          example: "3 Years"
        units:
          type: number
          example: 25
        shippingTime:
          type: string
          example: "2-3 business days"
        shippingCost:
          type: number
          example: 0
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                example: "https://example.com/image.jpg"
              thumb:
                type: string
                example: "https://example.com/image-thumb.jpg"
              isPrimary:
                type: boolean
        specifications:
          type: object
          additionalProperties:
            type: string
          example: {}
        inStock:
          type: boolean
        soldCount:
          type: number
        ratings:
          type: object
          properties:
            average:
              type: number
              example: 4.5
            count:
              type: number
              example: 128
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductCreate:
      type: object
      required:
        - name
        - heading
        - description
        - category
        - brand
        - price
        - cost
        - warranty
        - units
        - shippingTime
        - shippingCost
      properties:
        name:
          type: string
          example: "Intel Core i9-13900K"
        heading:
          type: string
          example: "24-Core Desktop Processor"
        description:
          type: string
          example: "The latest Intel processor with 24 cores and 32 threads"
        category:
          type: string
          example: "CAT488199006006"
          description: "Category ID (can be ObjectId, categoryId, or slug)"
        brand:
          type: string
          example: "Intel"
          description: "Product brand (required)"
        model:
          type: string
          example: "Core i9-13900K"
          description: "Product model (optional)"
        price:
          type: number
          example: 58999
        cost:
          type: number
          example: 52000
        warranty:
          type: string
          example: "3 Years"
        units:
          type: number
          example: 25
        shippingTime:
          type: string
          example: "2-3 business days"
        shippingCost:
          type: number
          example: 0
        specifications:
          type: object
          example: {
          }
        images:
          type: array
          items:
            type: string
            format: binary
          description: "Product images (max 5)"
...
    User:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
          example: "USR1234567890123456"
          description: "Auto-generated user ID (format: USR + timestamp + random)"
        name:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        cart:
          type: string
          description: "Cart ID reference"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Cart:
      type: object
      properties:
        _id:
          type: string
        cartId:
          type: string
          example: "CART1234567890123456"
          description: "Auto-generated cart ID"
        user:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: number
              price:
                type: number
              total:
                type: number
        subtotal:
          type: number
        shippingCost:
          type: number
        tax:
          type: number
        total:
          type: number
        coupon:
          type: object
          properties:
            code:
              type: string
            discount:
              type: number
        expiresAt:
          type: string
          format: date-time

    # ============= Layout Schemas =============
...
    IPRule:
      type: object
      properties:
        _id:
          type: string
        address:
          type: string
          example: "192.168.1.1"
        type:
          type: string
          enum: [whitelist, blacklist]
        reason:
          type: string
        addedBy:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ============= Search Schemas =============
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Not authorized to access this route
    ForbiddenError:
      description: User role not authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: User role user is not authorized to access this route
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              errors:
                type: array
                items:
                  type: object
                  properties:
                    msg:
                      type: string
                    param:
                      type: string
                    location:
                      type: string
    RecaptchaError:
      description: reCAPTCHA verification failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: reCAPTCHA verification failed
              error:
                type: string
                example: "reCAPTCHA score too low: 0.2"
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user with reCAPTCHA protection
      description: |
        Creates a new user account with auto-generated userId and cart.
        
        **Features:**
        - Auto-generates unique userId (format: USR + timestamp + random)
        - Auto-creates an empty cart for the user
        - Sends verification email
        - Optional address can be added later
        
        **reCAPTCHA Requirements:**
        - Token must be obtained from Google reCAPTCHA (v2 or v3)
        - For v3, score should be above configured threshold (default 0.5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              basic:
                summary: Basic signup without address
                value:
                  name: "John"
                  surname: "Doe"
                  email: "john.doe@example.com"
                  password: "Password123"
                  phone: "9876543210"
                  recaptchaToken: "03AGdBq27s_7xU3zQYxK5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X"
              withAddress:
                summary: Signup with address
                value:
                  name: "Jane"
                  surname: "Smith"
                  email: "jane.smith@example.com"
                  password: "Password123"
                  phone: "9876543211"
                  recaptchaToken: "03AGdBq27s_7xU3zQYxK5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X5X"
                  addresses:
                    - type: "primary"
                      name: "Jane Smith"
                      phone: "9876543211"
                      addressLine1: "456 Park Avenue"
                      city: "Delhi"
                      state: "Delhi"
                      pincode: "110001"
      responses:
        201:
          description: User registered successfully with auto-created cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Registration successful. Please check your email to verify your account.
        400:
          oneOf:
            - $ref: '#/components/responses/ValidationError'
            - $ref: '#/components/responses/RecaptchaError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user with reCAPTCHA protection
      description: |
        Authenticates a user and returns JWT token.
        
        **Features:**
        - Updates last login timestamp
        - Resets login attempts on success
        - Locks account after 5 failed attempts
        
        **reCAPTCHA Requirements:**
        - Token must be obtained from Google reCAPTCHA
        - Failed attempts with low scores may be rate limited
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/RecaptchaError'
        401:
          description: Invalid credentials or account not activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid credentials
                  needsActivation:
                    type: boolean
                    example: true
        423:
          description: Account locked
...
  /categories:
    get:
      tags: [Categories]
      summary: Get all categories as a tree structure with brands
      description: Returns categories in a hierarchical tree structure with brands at every level.
      responses:
        200:
          description: Category tree with brands at all levels
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: number
                    example: 10
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryTree'

    post:
      tags: [Categories]
      summary: Create a new category with brands (Admin only)
      description: |
        Creates a new category with auto-generated categoryId and slug.
        Brands can be provided as an array or comma-separated string.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
            examples:
              rootWithBrands:
                summary: Create root category with brands array
                value:
                  name: "Electronics"
                  description: "Electronic items and gadgets"
                  brands: ["Samsung", "Sony", "LG", "Panasonic"]
                  order: 1
              
              childWithBrandsString:
                summary: Create child category with comma-separated brands
                value:
                  name: "Laptops"
                  description: "Portable computers"
                  brands: "Dell,HP,Lenovo,Apple,Asus,Acer"
                  parent: "CAT123456781234"
                  order: 1
      responses:
        201:
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryTree'
        400:
          description: Validation error or parent not found
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
...
  /products:
    get:
      tags: [Products]
      summary: Get all products with filters
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category (ID, categoryId, or slug)
        - in: query
          name: brand
          schema:
            type: string
          description: Filter by brand name
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular, brand_asc]
        - in: query
          name: inStock
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
          description: Text search across name, description, brand, and model
      responses:
        200:
          description: List of products with brand and model information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  total:
                    type: number
                  page:
                    type: number
                  pages:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'

    post:
      tags: [Products]
      summary: Create a new product with brand and model (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        201:
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
...
  /users/orders:
    get:
      tags: [Users]
      summary: Get own orders
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of user orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
...
  /cart:
    get:
      tags: [Cart]
      summary: Get user cart (auto-creates if not exists)
      security:
        - bearerAuth: []
      responses:
        200:
          description: Cart data with products including brand and model
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Cart'

    delete:
      tags: [Cart]
      summary: Clear cart
      security:
        - bearerAuth: []
      responses:
        200:
          description: Cart cleared
...
  /orders:
    post:
      tags: [Orders]
      summary: Create new order from cart
      description: Creates an order from the user's cart and clears the cart.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shippingAddress
              properties:
                shippingAddress:
                  $ref: '#/components/schemas/Address'
                billingAddress:
                  $ref: '#/components/schemas/Address'
                paymentMethod:
                  type: string
                  enum: [razorpay]
                  default: razorpay
      responses:
        201:
          description: Order created with payment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
                      payment:
                        type: object
                        properties:
                          razorpayOrderId:
                            type: string
                          amount:
                            type: number
                          currency:
                            type: string
...
  /layout:
    get:
      tags: [Layout]
      summary: Get active layout with populated data
      responses:
        200:
          description: Complete layout with products and categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Layout'

    put:
      tags: [Layout]
      summary: Update layout (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutUpdate'
      responses:
        200:
          description: Layout updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      banner:
                        type: array
                      topItems:
                        type: array
                        items:
                          type: string
                      customSections:
                        type: array

  # ============= SEARCH ENDPOINTS =============
  /search:
    get:
      tags: [Search]
      summary: Search products across name, description, brand, and model
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: category
          schema:
            type: string
          description: Category slug to filter
        - in: query
          name: brand
          schema:
            type: string
          description: Filter by brand
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
      responses:
        200:
          description: Search results with suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  total:
                    type: number
                  page:
                    type: number
                  pages:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  suggestions:
                    type: array
                    items:
                      type: string
...
  /ip:
    get:
      tags: [IP]
      summary: Get all IP rules (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [whitelist, blacklist]
      responses:
        200:
          description: List of IP rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IPRule'

    post:
      tags: [IP]
      summary: Add IP rule (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
                - type
              properties:
                address:
                  type: string
                  example: "192.168.1.100"
                type:
                  type: string
                  enum: [whitelist, blacklist]
                reason:
                  type: string
                expiresAt:
                  type: string
                  format: date-time
      responses:
        201:
          description: IP rule created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/IPRule'
...